* CodeChallenge.Android.csproj: Feature 1: Show MovieDetailPage completed
* README.md:
* App.xaml:
* App.xaml.cs:
* Info.plist:
* Genre.cs:
* Movie.cs:
* AppDelegate.cs:
* HomePage.xaml:
* CodeChallenge.csproj:
* ITmdbApi.cs:
* HomePage.xaml.cs:
* MainActivity.cs:
* MovieService.cs:
* MovieItemView.xaml:
* Dates.cs:
* MovieDetailPage.xaml:
* Result.cs:
* MovieItemView.xaml.cs:
* CodeChallenge.iOS.csproj:
* MovieDetailPage.xaml.cs:
* HomePageViewModel.cs:
* MovieItemViewModel.cs:
* GenreResponse.cs:
* MovieDetailPageViewModel.cs:
* AndroidManifest.xml:
* UpcomingMoviesResponse.cs:
