+++ CodeChallenge.Android/CodeChallenge.Android.csproj
@@ -57,6 +57,9 @@
     <PackageReference Include="Refit">
       <Version>4.6.99</Version>
     </PackageReference>
+    <PackageReference Include="Xamarin.Forms.Visual.Material">
+      <Version>3.6.0.293080</Version>
+    </PackageReference>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="MainActivity.cs" />
