+++ CodeChallenge.iOS/Info.plist
@@ -24,15 +24,15 @@
 	<string>11.0</string>
 	<key>CFBundleDisplayName</key>
 	<string>CodeChallenge</string>
-	<key>CFBundleIdentifier</key>
-	<string>com.arctouch.codechallenge</string>
 	<key>CFBundleVersion</key>
 	<string>1.0</string>
 	<key>UILaunchStoryboardName</key>
 	<string>LaunchScreen</string>
-	<key>CFBundleName</key>
-	<string>Code Challenge</string>
 	<key>XSAppIconAssets</key>
 	<string>Assets.xcassets/AppIcon.appiconset</string>
+	<key>CFBundleIdentifier</key>
+	<string>dev.haroldjose.cinelovers</string>
+	<key>CFBundleName</key>
+	<string>CineLovers</string>
 </dict>
 </plist>
