+++ CodeChallenge/App.xaml.cs
@@ -33,7 +33,7 @@ namespace CodeChallenge
         {
             InitializeComponent();
 
-            MainPage = new HomePage();
+            MainPage = new NavigationPage(new HomePage());
         }
 
         protected override async void OnStart()
