+++ CodeChallenge/CodeChallenge.csproj
@@ -12,6 +12,7 @@
   <ItemGroup>
     <PackageReference Include="Xamarin.Forms" Version="3.6.0.293080" />
     <PackageReference Include="Refit" Version="4.6.99" />
+    <PackageReference Include="Xamarin.Forms.Visual.Material" Version="3.6.0.293080" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Services\" />
@@ -21,5 +22,6 @@
     <Folder Include="Utils\" />
     <Folder Include="Views\" />
     <Folder Include="Effects\" />
+    <Folder Include="Models\Responses\" />
   </ItemGroup>
 </Project>
