+++ CodeChallenge/Models/Genre.cs
@@ -20,80 +20,8 @@ using System.Collections.Generic;
 
 namespace CodeChallenge.Models
 {
-    public class UpcomingMoviesResponse
-    {
-        public List<Movie> Results { get; set; }
 
-        public int Page { get; set; }
 
-        public int TotalResults { get; set; }
-
-        public Dates Dates { get; set; }
-
-        public int TotalPages { get; set; }
-    }
-
-    public class GenreResponse
-    {
-        public List<Genre> Genres { get; set; }
-    }
-
-    public class Dates
-    {
-        public DateTimeOffset Maximum { get; set; }
-
-        public DateTimeOffset Minimum { get; set; }
-    }
-
-    public class Result
-    {
-        public int VoteCount { get; set; }
-
-        public int Id { get; set; }
-
-        public bool Video { get; set; }
-
-        public double VoteAverage { get; set; }
-
-        public string Title { get; set; }
-
-        public double Popularity { get; set; }
-
-        public string PosterPath { get; set; }
-
-        public string OriginalLanguage { get; set; }
-
-        public string OriginalTitle { get; set; }
-
-        public List<int> GenreIds { get; set; }
-
-        public string BackdropPath { get; set; }
-
-        public bool Adult { get; set; }
-
-        public string Overview { get; set; }
-
-        public DateTimeOffset ReleaseDate { get; set; }
-    }
-
-    public class Movie
-    {
-        public int Id { get; set; }
-
-        public string BackdropPath { get; set; }
-
-        public List<int> GenreIds { get; set; }
-
-        public List<Genre> Genres { get; set; }
-
-        public string Overview { get; set; }
-
-        public string PosterPath { get; set; }
-
-        public DateTimeOffset ReleaseDate { get; set; }
-
-        public string Title { get; set; }
-    }
 
     public class Genre
     {
