+++ CodeChallenge/Models/Responses/UpcomingMoviesResponse.cs
@@ -1,5 +1,5 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="AppDelegate.cs" company="ArcTouch LLC">
+// <copyright file="UpcomingMoviesResponse.cs" company="ArcTouch LLC">
 //   Copyright 2019 ArcTouch LLC.
 //   All rights reserved.
 //
@@ -11,29 +11,27 @@
 //   the license agreement.
 // </copyright>
 // <summary>
-//   Defines the AppDelegate type.
+//   Defines the UpcomingMoviesResponse type.
 // </summary>
 //  --------------------------------------------------------------------------------------------------------------------
+using System;
+using System.Collections.Generic;
 
-using Foundation;
-using UIKit;
-
-namespace CodeChallenge.iOS
+namespace CodeChallenge.Models.Responses
 {
-    [Register("AppDelegate")]
-    public partial class AppDelegate : global::Xamarin.Forms.Platform.iOS.FormsApplicationDelegate
+    public class UpcomingMoviesResponse
     {
-        public override bool FinishedLaunching(UIApplication app, NSDictionary options)
-        {
-            Xamarin.Calabash.Start();
-            global::Xamarin.Forms.Forms.Init();
+        public List<Movie> Results { get; set; }
 
-            //initialize Material design template from nuget
-            global::Xamarin.Forms.FormsMaterial.Init();
+        public int Page { get; set; }
 
-            LoadApplication(new App());
+        public int TotalResults { get; set; }
 
-            return base.FinishedLaunching(app, options);
-        }
+        public Dates Dates { get; set; }
+
+        public int TotalPages { get; set; }
     }
+
+
+
 }
