+++ CodeChallenge/Views/MovieDetailPage.xaml
@@ -1,0 +1,51 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms" 
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" 
+             xmlns:d="http://xamarin.com/schemas/2014/forms/design"
+             x:Class="CodeChallenge.Views.MovieDetailPage"
+             Visual= "Material">
+    <ContentPage.Content>
+        <ScrollView>
+            <StackLayout>
+                <Image
+                    Aspect="AspectFit"
+                    Source="{Binding PosterPath}"/>
+                <Label
+                    FontSize="30"
+                    FontAttributes="Bold"
+                    TextColor="Black"
+                    HorizontalTextAlignment="Start"
+                    Text="{Binding Title}"/>
+                <Label
+                    Grid.Row="1"
+                    Grid.Column="1"
+                    FontSize="Small"
+                    TextColor="Black"
+                    HorizontalTextAlignment="Start"
+                    Text="{Binding Genres}"/>
+                 <Label
+                    Grid.Row="2"
+                    Grid.Column="1"
+                    FontSize="Micro"
+                    FontAttributes="Italic"
+                    TextColor="Black"
+                    VerticalTextAlignment="End"
+                    Text="{Binding ReleaseDate, StringFormat='{}{0:dd/MM/yyyy}'}"
+                    />
+                <Label
+                    Grid.Row="2"
+                    Grid.Column="1"
+                    FontSize="Micro"
+                    FontAttributes="Italic"
+                    TextColor="Black"
+                    VerticalTextAlignment="End"
+                    Text="{Binding Overview}"/>
+                <Image
+                    Aspect="AspectFit"
+                    Source="{Binding BackdropPath}"/>
+            </StackLayout>    
+        </ScrollView>
+    </ContentPage.Content>
+</ContentPage>
+
+
