+++ CodeChallenge/Views/MovieItemView.xaml
@@ -4,12 +4,17 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:effects="clr-namespace:CodeChallenge.Effects"
     x:Class="CodeChallenge.Views.MovieItemView">
-    <Frame CornerRadius="{x:OnPlatform Android=0, iOS=3}"
+     <Frame CornerRadius="{x:OnPlatform Android=0, iOS=3}"
            HasShadow="{x:OnPlatform Android=true, iOS=false}"
            BorderColor="{x:OnPlatform Android=Transparent, iOS=Gray}"
            Padding="{x:OnPlatform Android=0, iOS=6}"
            Margin="4">
         <Grid RowSpacing="0">
+            
+             <Grid.GestureRecognizers>
+                    <TapGestureRecognizer Command="{Binding TapCommand}" />
+                </Grid.GestureRecognizers>       
+            
             <Grid.RowDefinitions>
                 <RowDefinition Height="Auto" />
                 <RowDefinition Height="Auto" />
@@ -22,7 +27,8 @@
             <Image
                 Grid.RowSpan="3"
                 Aspect="AspectFit"
-                Source="{Binding PosterPath}"/>
+                Source="{Binding PosterPath}">
+            </Image>
             <Label
                 Grid.Row="0"
                 Grid.Column="1"
