+++ README.md
@@ -1,0 +1,9 @@
+
+
+
+Componentes Used:
+- Xamarin.Forms.Visual.Material
+
+
+Design Inspirations
+- https://www.behance.net/gallery/61091395/Moveee-Movie-app-concept
